requirements.txt
openai==0.27.0  # Versão compatível com a API da OpenAI
streamlit==1.24.0  # Versão para o front-end
fpdf==1.7.2 # Versão para a biblioteca que salva PDF
python-dotenv==1.0.1